MetaHS.DataModel.MetaModel

Description

MetaHS metamodel

Documentation

data Element

The various elements that can be used in the metamodel.

Constructors

Program
Top level program

Fields

name :: !String

The name of the program.

Module
Represents a module

Fields

name :: !String

The name of the program.

ModuleHead
Represents a module head

Fields

name :: !String

The name of the program.

ModuleImport
Represents a module import

Fields

name :: !String

The name of the program.

ModuleExport
Represents a module export

Fields

name :: !String

The name of the program.

Function
Represents a function

Fields

name :: !String

The name of the program.

DataType
Represents a datatype declaration.

Fields

name :: !String

The name of the program.

TypeSynonym
Represents a type synonym.

Fields

name :: !String

The name of the program.

TypeSignature
Represents a type signature.

Fields

name :: !String

The name of the program.

TypeClass
Represents a type instance.

Fields

name :: !String

The name of the program.

Instance

Fields

name :: !String

The name of the program.

Pragma
Represents an inline Pragma.

Fields

name :: !String

The name of the program.

UnknownType
Represents a type which is unknown.

Fields

name :: !String

The name of the program.

Location
Represents a source code location.

Fields

locationPath :: !String

The path to the file containing the source code.

locationStartLine :: !Int

The start line.

locationStartColumn :: !Int

The start column.

locationEndLine :: !Int

The end line.

locationEndColumn :: !Int

The end column.

StringValue
Contains a generic String value.

Fields

stringValue :: !String

The String value.

IntValue
Contains a generic Int value.

Fields

intValue :: !Int

The Int value.

type Pair = (Element, Element)

A Pair is defined as a tuple of Elements

type Relation = Set (Element, Element)

A relation is defined as a set of Pair types. E.g., _contains, _source, _uses, LCOM

newtype MetaModel

The MetaModel type.

Constructors

MetaModel

Fields

getMetaModelImpl :: MetaModelImpl

pPrint :: Pretty a => a -> Doc