MetaHS.DataModel.Utils.Find

Description

Generic SYB functions for finding certain nodes

Documentation

findTyConNames

Arguments

:: Data a

=> a
The data structure to search.

-> [String]
The names of the TyCon nodes found.

SYB function that searches for TyCon nodes and returns their names.

findQNames

Arguments

:: Data a

=> a
The data structure to search.

-> [String]
The names of the QNames nodes found.

SYB function that searches for QName nodes (without TyCons) and returns their names.

findPatVars

Arguments

:: Data a

=> a
The data structure to search.

-> [String]
The names of the Pat variables found.

SYB function that searches for Pat nodes that represent variables and returns their names.

findFunctionNames

Arguments

:: Data a

=> a
The data structure to search.

-> [String]
The function names found.

SYB function that searches for functions and returns their names.