MetaHS.DataModel.Utils.Language.Haskell.Exts.SrcLoc

Description

Utility functions for SrcLoc and related objects.

Documentation

srcLocToLocationElement

Arguments

:: SrcLoc
The SrcLoc object to analyze.

-> Element
The resulting Element.

Creates a Location Element containing the information of the SrcLoc object.

srcSpanToLocationElement

Arguments

:: SrcSpan
The SrcSpan object to analyze.

-> Element
The resulting Element.

Creates a Location Element containing the information of the SrcSpan object.

srcSpanInfoToLocationElement

Arguments

:: SrcSpanInfo
The SrcSpanInfo object to analyze.

-> Element
The resulting Element.

Creates a Location Element containing the information of the SrcSpanInfo object.