MetaHS.DataModel.Utils.Language.Haskell.Exts.Syntax.Decl

Description

Utility functions for Decl objects.

Documentation

srcSpanInfo :: Decl SrcSpanInfo -> SrcSpanInfo

Returns the SrcSpanInfo object associated with the Decl object.

data DataConstructor

Represents the (simplified) information for a data constructor.

Constructors

DataConstructor

Fields

dataConstructorName :: String

The name of the data constructor.

valueConstructors :: [ValueConstructor]

The value constructors for this data constructor.

data ValueConstructor

Represents the information for a value constructor.

Constructors

ValueConstructor

Fields

valueConstructorName :: String

The name of the value constructor.

valueConstructorTypes :: [Type SrcSpanInfo]

Contains the types used by this value constructor if this is not defined using record notation. For GADT data type declarations, this field will contain the final type irrespective of whether record notation is used for the preceeding types.

valueConstructorFields :: [Field]

In case the value constructor is defined using record notation, this field will contain the fields used by this value constructor.

data Field

Represents a record field of a value constructor.

Constructors

Field

Fields

fieldNames :: [String]

The name(s) for this field.

fieldTypes :: Type SrcSpanInfo

The types used by this field.

dataConstructor :: Decl SrcSpanInfo -> Maybe DataConstructor

Converts a Decl to a DataConstructor object if possible (DataDecl or GDataDecl).

patternName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the PatBind if possible.

functionName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the FunBind if possible.

typeSigName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the TypeSig if possible.

typeClassName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the typeClass if possible.

instanceName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the Instance (as "type-class data-type") if possible.

inlineSigName :: Decl SrcSpanInfo -> Maybe String

Returns the name of the InlineSig if possible.