MetaHS.DataModel.Utils.NameResolution

Description

NameResolution functions

Documentation

type NameResolutionMap = Map String Element

A NameResolutionMap is a Data.Map that map from plain identifiers to resolved MetaModel.Element (e.g. "sum" -> Function "M.N.sum").

type NameResolutionMaps = (NameResolutionMap, NameResolutionMap)

The NameResolutionMaps. The first Data.Map is applicable to types; the second Data.Map is applicable to values.

createNameResolutionMaps

Arguments

:: String
Module name.

-> Relation
The "Module mn Contains Element" Relations.

-> NameResolutionMaps
The resulting NameResolutionMaps.

Creates the NameResolution Data.Map instances.

resolveType

Arguments

:: String
The simple value name.

-> NameResolutionMaps
The NameResolutionMaps.

-> Element
The corresponding Element.

Resolves a simple type name to the corresponding qualified Element. The simple type name will be qualified with "?" and wrapped in an MetaModel.UnknownType Element if the mapping is not known by the provided NameResolutionMaps.

resolveValue

Arguments

:: String
The simple value name.

-> NameResolutionMaps
The NameResolutionMaps.

-> Element
The corresponding Element.

Resolves a simple value name to the corresponding qualified Element. The simple value name will be qualified with "?" if the mapping is not known by the provided NameResolutionMaps.