MetaHS.EDSL.Graph.UsesGraph

Description

Generates a Uses graph for a module

Documentation

internalUses

Arguments

:: MetaModel
The meta-model.

-> Element
The module.

-> Directed
This option will determine whether the resulting graph is directed or undirected.

-> String
The prefix for the links in the generated SVG images to the correct HTML editor for displaying the source code. Should probably become a hard link if a web server is used.

-> (GraphType, ParamsType)
The generated Graph.

Generates the internal uses graph for the specified module and returns the associated default GraphvizParams.

internalUsesGraph

Arguments

:: MetaModel
The meta-model.

-> Element
The module.

-> Directed
This option will determine whether the resulting graph is directed or undirected.

-> GraphType
The generated Graph.

Generates the internal uses graph for the specified module.

internalUsesParams

Arguments

:: MetaModel
The meta-model.

-> Element
The module.

-> Directed
This option will determine whether the resulting graph is directed or undirected.

-> String
The prefix for the links in the generated SVG images to the correct HTML editor for display the source code. Should probably become a hard link if a web server is used.

-> ParamsType
The generated Graph.

Generates the default GraphvizParams for the internal uses graph for the specified module.