MetaHS.EDSL.Graph.Utils

Description

Utilities for handling graphs

Documentation

graphToImage

Arguments

:: GraphvizCommand
The Graphviz layouter to use.

-> GraphvizOutput
The output format.

-> FilePath
The FilePath to use for the generated image.

-> GraphType
The graph to render.

-> ParamsType
The graph parameters to use.

-> IO FilePath
The FilePath used for the generated image.

Generates an image from a graph and writes this to a file in the specified format using the specified Graphviz layouter.