MetaHS.EDSL.MetaModel

Description

MetaHS EDSL MetaModel part

Documentation

generateMetaModel

Arguments

:: String
The program name.

-> String
The program path.

-> String
The parse error output path.

-> IO MetaModel
The resulting meta-model.

Generates a meta-model.

writeMetaModel

Arguments

:: MetaModel
The meta-model to write.

-> String
The path of the file to write.

-> IO ()
No result.

Write a meta-model to a file.

readMetaModel

Arguments

:: String
The path of the file to read.

-> IO MetaModel
The meta-model read.

Read a meta-model from a file.

writeMetaModelPretty

Arguments

:: MetaModel
The meta-model to write.

-> String
The path of the file to write.

-> IO ()
No result.

Write a meta-model to a file.

pretty

Arguments

:: MetaModel
The meta-model to pretty print.

-> Doc
The resulting Doc.

Returns a pretty printed representation of the meta-model.

numberOfItems

Arguments

:: MetaModel
The meta-model.

-> Int
The number of relation items in the meta-model.

Returns the number of relation items in the meta-model.

getPrograms

Arguments

:: MetaModel
The metamodel.

-> [Element]
The Programs contained in the metamodel.

Returns a list of Programs contained in the metamodel.

getModules

Arguments

:: MetaModel
The metamodel.

-> [Element]
The Modules contained by the specified Program.

Returns a list of Modules contained in the metamodel.

programContains

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Program.

-> [Element]
The Elements contained by the specified Program.

Returns a list of Elements contained by the specified Program.

moduleContains

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Module.

-> [Element]
The Elements contained by the specified Module.

Returns a list of Elements contained by the specified Module.

moduleImports

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Module.

-> [Element]
The Elements contained by the specified Module.

Returns a list of Elements imported by the specified Module.

elementContains

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Element.

-> [Element]
The Elements contained by the specified Element.

Returns a list of Elements contained by the specified Element.

elementImports

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Element.

-> [Element]
The Elements contained by the specified Element.

Returns a list of Elements imported by the specified Element.

elementSource

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Element.

-> Maybe Element
The source location for the specified Element.

Returns a the source location for the specified Element.

elementUses

Arguments

:: MetaModel
The meta-model.

-> Element
The specified Element.

-> [Element]
The Elements used by the specified Element.

Returns a list of Elements used by the specified Element.

getRelation

Arguments

:: String
The metamodel key string.

-> MetaModel
The metamodel.

-> Relation
The corresponding Relation or an empty set.

Returns the relation corresponding to the supplied key or an empty Set.

setRelation

Arguments

:: String
The metamodel key string.

-> Relation
The Relation to add.

-> MetaModel
The metamodel.

-> MetaModel
The resulting metamodel.

Returns a MetaModel that is appended with a new or replaced Relation set.

domain

Arguments

:: String
The metamodel key string.

-> MetaModel
The metamodel.

-> Set Element
Set of Elements that form the domain of the binary relation.

Returns the domain for a specified relation in the metamodel.

range

Arguments

:: String
The metamodel key string.

-> MetaModel
The metamodel.

-> Set Element
Set of Elements that form the range of the binary relation.

Returns the range for a specified relation in the metamodel.

type RelationKey = String

convenience type that represents the metamodel aggregation relation