MetaHS.EDSL.Utils

Description

MetaHS EDSL Utils part

Documentation

split

Arguments

:: String
The qualified String to split.

-> (String, String)
The qualified and name parts.

Splits an identifier in the qualifier and name parts.

isLocal

Arguments

:: String
The qualifier.

-> String
The qualified name.

-> Bool
True if the qualified name is local to the qualifier.

Determines whether a given qualified name is local to the given qualifier.

locationToQuery

Arguments

:: Element
The Location Element.

-> String
The resulting URL query String or an empty String.

Converts a Location Element to a query String suitable for URLs. An empty String is returned if the Element is not a Location Element.