MetaHS.Extensions.CBO

Description

Metamodel Element that is interesting (those that represents modules being imported) is : Module. All other Elements that are part of the _contains relations are ignored.

CBO - Coupling between object classes The coupling between object classes (CBO) metric represents the number of classes coupled to a given class (efferent couplings, Ce). This coupling can occur through method calls, field accesses, inheritance, arguments, return types, and exceptions. ref. Chidamber and Kemerer (1994)

Documentation

keyCbo :: RelationKey

MetaModel key used for the CBO relation.

cboAggregator

Arguments

:: MetaModel
The supplied meta-model.

-> MetaModel
The complemented meta-model.

CBO aggregator that adds the CBO metric relation to the supplied meta-model.