MetaHS.Extensions.LCOM

Description

Metamodel Elements that are interesting (those that are used to determine nodes in MetaHS.EDSL.Graph.UsesGraph module ) are : TypeSynonym, DataType, Function. All other Elements that are part of the _contains relations are ignored.

LCOM4 - Number of connected components in the graph that represents each method as a node and the sharing of at least one attribute/method as an edge. ref. Hitz and Montazeri (1995)

Documentation

lcom

Arguments

:: MetaModel
The meta-model.

-> Element
The Module Element.

-> Int
The calculated LCOM metric value.

Calculates the LCOM metric value for a specified module name.

lcomGraph

Arguments

:: MetaModel
The meta-model.

-> Element
The Module Element.

-> Directed
Create a directed or undirected graph.

-> String
The prefix for the links in the generated SVG images to the correct HTML editor for display the source code. Should probably become a hard link if a web server is used.

-> (Int, GraphType, ParamsType)
The calculated LCOM metric value and the generated Graph and default parameters.

Generates the internalUsesGraph for the specified Module Element and calculates the corresponding LCOM metric value.

keyLcom :: RelationKey

MetaModel key used for the LCOM relation.

lcomAggregator :: MetaModel -> MetaModel

LCOM aggregator that adds the LCOM relation to the metamodel. key = keyLcom value = (p,c) where p is a Module Element and c is an IntValue Element that contains the LCOM value for the module.