MetaHS.Extensions.MacroLevelAggregation.GiniCoefficient

Description

Calculates the Gini-coefficient of the metric population corresponding to supplied Relation Key. Result is a value bounded by [0,1].

Documentation

giniCoefficient

Arguments

:: RelationKey
The MetaModel Relation Key (E.g., LCOM).

-> MetaModel
The MetaModel Containing the associated key.

-> Double
The Gini coefficient

Calculate the Gini-coefficient of the metric values associated with the supplied metric Relation key